test(protocol): share sequence test helper
Refs #1342
